rename: ErrorMessage를 error 패키지로 이동
