+++ src/main/java/lotto/error/ErrorMessage.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.error;
 
 public enum ErrorMessage {
